Updated components definitions
