+++ packages/types/src/components/index.ts
@@ -1,5 +1,1 @@
 export type * from "./types";
-export type * from "./box";
-export type * from "./col";
-export type * from "./row";
-export type * from "./field";
